backend/requirements.txt
Flask
flask-cors
python-dotenv
psycopg2-binary
python-dateutil
pandas
numpy
joblib
